Improve tests for SlackNotifier
