+++ tests/Unit/ScrumMaster/Command/SlackNotifierCommandTest.php
@@ -7,17 +7,17 @@ namespace App\Tests\Unit\ScrumMaster\Command;
 use App\ScrumMaster\Command\SlackNotifierCommand;
 use App\ScrumMaster\Command\SlackNotifierInput;
 use App\ScrumMaster\Jira\JiraHttpClient;
-use App\ScrumMaster\Jira\Tickets;
 use App\ScrumMaster\Slack\SlackHttpClient;
-use PHPUnit\Framework\MockObject\MockObject;
+use App\Tests\Unit\ScrumMaster\JiraApiResource;
 use PHPUnit\Framework\TestCase;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
-use Symfony\Contracts\HttpClient\ResponseInterface;
 
 final class SlackNotifierCommandTest extends TestCase
 {
+    use JiraApiResource;
+
     /** @test */
-    public function zeroSuccessfulNotificationsWereSent(): void
+    public function zeroNotificationsWereSent(): void
     {
         $output = new InMemoryOutput();
 
@@ -44,10 +44,12 @@ final class SlackNotifierCommandTest extends TestCase
     public function twoSuccessfulNotificationsWereSent(): void
     {
         $output = new InMemoryOutput();
+
         $jiraIssues = [
             $this->createAnIssueAsArray('user.1.jira'),
             $this->createAnIssueAsArray('user.2.jira'),
         ];
+
         $command = new SlackNotifierCommand(
             new JiraHttpClient($this->mockJiraClient($jiraIssues)),
             new SlackHttpClient($this->createMock(HttpClientInterface::class))
@@ -60,39 +62,6 @@ final class SlackNotifierCommandTest extends TestCase
             'SLACK_MAPPING_IDS' => '{"jira.id":"slack.id"}',
         ]), $output);
 
-        $this->assertEquals('Total notifications: 2', $output->lines()[0]);
-    }
-
-    private function mockJiraClient(array $issues): HttpClientInterface
-    {
-        $jiraResponse = $this->createMock(ResponseInterface::class);
-        $jiraResponse->method('toArray')->willReturn(['issues' => $issues]);
-
-        /** @var HttpClientInterface|MockObject $jiraClient */
-        $jiraClient = $this->createMock(HttpClientInterface::class);
-        $jiraClient->method('request')->willReturn($jiraResponse);
-
-        return $jiraClient;
-    }
-
-    private function createAnIssueAsArray(string $assigneeName): array
-    {
-        return [
-            'key' => 'KEY-123',
-            'fields' => [
-                Tickets::FIELD_STORY_POINTS => '5.0',
-                'status' => [
-                    'name' => 'In Progress',
-                ],
-                'summary' => 'The ticket title',
-                'statuscategorychangedate' => '2019-06-15T10:35:00+00',
-                'assignee' => [
-                    'name' => $assigneeName,
-                    'key' => 'user.key.jira',
-                    'emailAddress' => 'user@email.jira',
-                    'displayName' => 'display.name.jira',
-                ],
-            ],
-        ];
+        $this->assertContains('Total notifications: 2', $output->lines());
     }
 }
